[UPGRADE] end refacto
