+++ nbactions.xml
@@ -10,7 +10,7 @@
         </goals>
         <properties>
             <spring-boot.run.jvmArguments>-noverify -XX:TieredStopAtLevel=1 </spring-boot.run.jvmArguments>
-            <spring-boot.run.mainClass>com.web1.demo.DemoApplication</spring-boot.run.mainClass>
+            <spring-boot.run.mainClass>com.medhead.api.RunApi</spring-boot.run.mainClass>
             <Env.SPRING_OUTPUT_ANSI_ENABLED>always</Env.SPRING_OUTPUT_ANSI_ENABLED>
         </properties>
     </action>
@@ -24,7 +24,7 @@
         </goals>
         <properties>
             <spring-boot.run.jvmArguments>-Xdebug -Xrunjdwp:transport=dt_socket,server=n,address=${jpda.address} -noverify -XX:TieredStopAtLevel=1 </spring-boot.run.jvmArguments>
-            <spring-boot.run.mainClass>com.web1.demo.DemoApplication</spring-boot.run.mainClass>
+            <spring-boot.run.mainClass>com.medhead.api.RunApi</spring-boot.run.mainClass>
             <Env.SPRING_OUTPUT_ANSI_ENABLED>always</Env.SPRING_OUTPUT_ANSI_ENABLED>
             <jpda.listen>true</jpda.listen>
         </properties>
@@ -39,7 +39,37 @@
             <goal>org.codehaus.mojo:exec-maven-plugin:1.2.1:exec</goal>
         </goals>
         <properties>
-            <exec.args>-classpath %classpath com.web1.demo.DemoApplication</exec.args>
+            <exec.args>-classpath %classpath com.medhead.api.RunApi</exec.args>
+            <spring-boot.run.mainClass>com.medhead.api.RunApi</spring-boot.run.mainClass>
+            <Env.SPRING_OUTPUT_ANSI_ENABLED>always</Env.SPRING_OUTPUT_ANSI_ENABLED>
+        </properties>
+    </action>
+    <action>
+        <actionName>debug</actionName>
+        <packagings>
+            <packaging>jar</packaging>
+        </packagings>
+        <goals>
+            <goal>spring-boot:run</goal>
+        </goals>
+        <properties>
+            <spring-boot.run.jvmArguments>-Xdebug -Xrunjdwp:transport=dt_socket,server=n,address=${jpda.address} -noverify -XX:TieredStopAtLevel=1 </spring-boot.run.jvmArguments>
+            <spring-boot.run.mainClass>com.medhead.api.RunApi</spring-boot.run.mainClass>
+            <Env.SPRING_OUTPUT_ANSI_ENABLED>always</Env.SPRING_OUTPUT_ANSI_ENABLED>
+            <jpda.listen>true</jpda.listen>
+        </properties>
+    </action>
+    <action>
+        <actionName>profile</actionName>
+        <packagings>
+            <packaging>jar</packaging>
+        </packagings>
+        <goals>
+            <goal>process-classes</goal>
+            <goal>org.codehaus.mojo:exec-maven-plugin:1.2.1:exec</goal>
+        </goals>
+        <properties>
+            <exec.args>-classpath %classpath com.medhead.api.RunApi</exec.args>
             <exec.executable>java</exec.executable>
         </properties>
     </action>
