+++ pom.xml
@@ -8,11 +8,11 @@
 		<version>2.7.0</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>com.web1</groupId>
-	<artifactId>demo</artifactId>
+	<groupId>com.medhead</groupId>
+	<artifactId>pocApiMedHead</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>demo</name>
-	<description>Demo project for Spring Boot</description>
+	<name>pocApiMedHead</name>
+	<description>POC Api Medhead</description>
 	<properties>
 		<java.version>11</java.version>
 	</properties>
