+++ src/main/java/com/medhead/api/exeptions/NotYetImplemented.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package com.web1.demo.exeptions;
+package com.medhead.api.exeptions;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
@@ -11,9 +11,9 @@ import org.springframework.web.bind.annotation.ResponseStatus;
  *
  * @author stiven
  */
-@ResponseStatus(value = HttpStatus.NOT_FOUND)
-public class NoHospitalFound extends RuntimeException{
-    public NoHospitalFound(String message) {
+@ResponseStatus(HttpStatus.NOT_IMPLEMENTED)
+public class NotYetImplemented extends Exception {
+  public NotYetImplemented(String message) {
       super(message);
   }
 }
