+++ src/main/java/com/medhead/api/model/Hospital.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package com.web1.demo.model;
+package com.medhead.api.model;
 import lombok.Data;
 
 import javax.persistence.*;
