+++ src/main/java/com/medhead/api/repository/HospitalRepository.java
@@ -2,12 +2,12 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Interface.java to edit this template
  */
-package com.web1.demo.repository;
+package com.medhead.api.repository;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import com.web1.demo.model.Hospital;
+import com.medhead.api.model.Hospital;
 /**
  *
  * @author stiven
