+++ src/main/resources/application.properties
@@ -1,5 +1,5 @@
 #Global configuration
-spring.application.name=web2
+spring.application.name=pocApiMedHead
 
  #Tomcat configuration
  server.port=9000
@@ -12,10 +12,6 @@ logging.level.org.springframework.boot.web.embedded.tomcat=INFO
 
 #Exception message
 server.error.include-message=always
-
-#API project osrm : https://github.com/Project-OSRM/osrm-backend/blob/master/docs/http.md#nearest-service
-#UTILISER PLUTOT : https://nominatim.openstreetmap.org/search.php?q=hospitals&viewbox=-3.91663%2C55.48027%2C-2.97935%2C52.23116&bounded=1&countrycodes=gb&format=jsonv2
-com.etcomment.web2.webApi=http://router.project-osrm.org/route/v1/driving/
 
 #spring.autoconfigure.exclude=org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration
